# Handle more scenarios for loop cloning