--- a/src/coreclr/jit/loopcloning.cpp
+++ b/src/coreclr/jit/loopcloning.cpp
@@ -1024,17 +1024,13 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
 
     LoopDsc*                         loop     = &optLoopTable[loopNum];
     JitExpandArrayStack<LcOptInfo*>* optInfos = context->GetLoopOptInfo(loopNum);
+    bool                             isIncreasingLoop = GenTree::StaticOperIs(loop->lpTestOper(), GT_LT, GT_LE);
 
-    if (GenTree::StaticOperIs(loop->lpTestOper(), GT_LT, GT_LE))
+    if (GenTree::StaticOperIs(loop->lpTestOper(), GT_LT, GT_LE, GT_GT, GT_GE))
     {
-        // Stride conditions
-        if (loop->lpIterConst() <= 0)
-        {
-            JITDUMP("> Stride %d is invalid\n", loop->lpIterConst());
-            return false;
-        }
+        LC_Ident ident;
 
-        // Init conditions
+       // Init conditions
         if (loop->lpFlags & LPFLG_CONST_INIT)
         {
             // Only allowing non-negative const init at this time.
@@ -1044,6 +1040,12 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
             {
                 JITDUMP("> Init %d is invalid\n", loop->lpConstInit);
                 return false;
+            }
+
+            if (!isIncreasingLoop)
+            {
+                // For decreasing loop, the init value needs to be checked against the array length
+                ident = LC_Ident(static_cast<unsigned>(loop->lpConstInit), LC_Ident::Const);
             }
         }
         else
@@ -1056,13 +1058,22 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
                 return false;
             }
 
-            LC_Condition geZero(GT_GE, LC_Expr(LC_Ident(initLcl, LC_Ident::Var)),
-                                LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            LC_Condition geZero;
+            if (isIncreasingLoop)
+            {
+                geZero = LC_Condition(GT_GE, LC_Expr(LC_Ident(initLcl, LC_Ident::Var)),
+                                      LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            }
+            else
+            {
+                // For decreasing loop, the init value needs to be checked against the array length
+                ident = LC_Ident(initLcl, LC_Ident::Var);
+                geZero = LC_Condition(GT_GE, LC_Expr(ident), LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            }
             context->EnsureConditions(loopNum)->Push(geZero);
         }
 
         // Limit Conditions
-        LC_Ident ident;
         if (loop->lpFlags & LPFLG_CONST_LIMIT)
         {
             int limit = loop->lpConstLimit();
@@ -1071,7 +1082,12 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
                 JITDUMP("> limit %d is invalid\n", limit);
                 return false;
             }
-            ident = LC_Ident(static_cast<unsigned>(limit), LC_Ident::Const);
+
+            if (isIncreasingLoop)
+            {
+                // For increasing loop, thelimit value needs to be checked against the array length
+                ident = LC_Ident(static_cast<unsigned>(limit), LC_Ident::Const);
+            }
         }
         else if (loop->lpFlags & LPFLG_VAR_LIMIT)
         {
@@ -1082,8 +1098,18 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
                 return false;
             }
 
-            ident = LC_Ident(limitLcl, LC_Ident::Var);
-            LC_Condition geZero(GT_GE, LC_Expr(ident), LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            LC_Condition geZero;
+            if (isIncreasingLoop)
+            {
+                // For increasing loop, thelimit value needs to be checked against the array length
+                ident  = LC_Ident(limitLcl, LC_Ident::Var);
+                geZero = LC_Condition(GT_GE, LC_Expr(ident), LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            }
+            else
+            {
+                geZero = LC_Condition(GT_GE, LC_Expr(LC_Ident(limitLcl, LC_Ident::Var)),
+                                      LC_Expr(LC_Ident(0, LC_Ident::Const)));
+            }
 
             context->EnsureConditions(loopNum)->Push(geZero);
         }
@@ -1107,15 +1133,22 @@ bool Compiler::optDeriveLoopCloningConditions(unsigned loopNum, LoopCloneContext
             return false;
         }
 
-        // GT_LT loop test: limit <= arrLen
-        // GT_LE loop test: limit < arrLen
+        // Increasing loops
+        // GT_LT loop test: (start < end) ==> (end <= start)
+        // GT_LE loop test: (start <= end) ==> (end < start)
+        //
+        // Decreasing loops
+        // GT_GT loop test: (end > start) ==> (end <= start)
+        // GT_GE loop test: (end >= start) ==> (end < start)
         genTreeOps opLimitCondition;
         switch (loop->lpTestOper())
         {
             case GT_LT:
+            case GT_GT:
                 opLimitCondition = GT_LE;
                 break;
             case GT_LE:
+            case GT_GE:
                 opLimitCondition = GT_LT;
                 break;
             default:
@@ -1603,13 +1636,6 @@ bool Compiler::optIsLoopClonable(unsigned loopInd)
         return false;
     }
 
-    // TODO-CQ: CLONE: Mark increasing or decreasing loops.
-    if ((loop.lpIterOper() != GT_ADD) || (loop.lpIterConst() != 1))
-    {
-        JITDUMP("Loop cloning: rejecting loop " FMT_LP ". Loop iteration operator not matching.\n", loopInd);
-        return false;
-    }
-
     if ((loop.lpFlags & LPFLG_CONST_LIMIT) == 0 && (loop.lpFlags & LPFLG_VAR_LIMIT) == 0 &&
         (loop.lpFlags & LPFLG_ARRLEN_LIMIT) == 0)
     {
@@ -1618,8 +1644,24 @@ bool Compiler::optIsLoopClonable(unsigned loopInd)
         return false;
     }
 
-    if (!((GenTree::StaticOperIs(loop.lpTestOper(), GT_LT, GT_LE) && (loop.lpIterOper() == GT_ADD)) ||
-          (GenTree::StaticOperIs(loop.lpTestOper(), GT_GT, GT_GE) && (loop.lpIterOper() == GT_SUB))))
+    bool isLessThanLimitCheck = GenTree::StaticOperIs(loop.lpTestOper(), GT_LT, GT_LE);
+    bool isGreaterThanLimitCheck = GenTree::StaticOperIs(loop.lpTestOper(), GT_GT, GT_GE);
+
+    // Increasing loop is the one that has "+=" increament operation and "< or <=" limit check.
+    bool isIncreasingLoop = (isLessThanLimitCheck && (loop.lpIterOper() == GT_ADD));
+
+    // Increasing loop is the one that has "-=" decrement operation and "> or >=" limit check. If the operation is "+=",
+    // make sure the constant is negative to give an effect of decrementing the iterator.
+    bool isDecreasingLoop = (isGreaterThanLimitCheck && ((loop.lpIterOper() == GT_SUB) ||
+                                                         ((loop.lpIterOper() == GT_ADD) && (loop.lpIterConst() < 0))));
+
+    //TODO-CQ: Handle other loops like:
+    // - The ones whose limit operator is "==" or "!="
+    // - The incrementing operator is multiple and divide
+    // - The ones that are inverted are not handled here for cases like "i *= 2" because
+    //   they are converted to "i + i".
+    //bool isOtherLoop = (loop.lpIterOper() == GT_DIV) || (loop.lpIterOper() == GT_MUL);
+    if (!(isIncreasingLoop || isDecreasingLoop))
     {
         JITDUMP("Loop cloning: rejecting loop " FMT_LP
                 ". Loop test (%s) doesn't agree with the direction (%s) of the loop.\n",