--- a/src/tests/JIT/Directed/Arrays/LoopCloning.csproj
+++ b/src/tests/JIT/Directed/Arrays/LoopCloning.csproj
@@ -1,0 +1,13 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <CLRTestPriority>1</CLRTestPriority>
+  </PropertyGroup>
+  <PropertyGroup>
+    <DebugType>PdbOnly</DebugType>
+    <AllowUnsafeBlocks>True</AllowUnsafeBlocks>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="LoopCloning.cs" />
+  </ItemGroup>
+</Project>