- Increasing loops that are incremented by more than 1 like "i += 2"


[TypeHashingAlgorithm.cs](https://github.com/dotnet/runtime/blob/main/src/coreclr/tools/Common/TypeSystem/Common/TypeHashingAlgorithms.cs)



https://github.com/dotnet/runtime/blob/7bac4e8fa611eb02ac1bd27b97a8570cbf321437/src/coreclr/tools/Common/TypeSystem/Common/TypeHashingAlgorithms.cs#L44-L49


- Decreasing loops like "i--", "i -= 2", etc.

https://github.com/dotnet/runtime/blob/7bac4e8fa611eb02ac1bd27b97a8570cbf321437/src/libraries/System.Private.CoreLib/src/System/Collections/Generic/EqualityComparer.cs#L50-L63



PriorityQueue's [Heapify](https://github.com/dotnet/runtime/blob/7bac4e8fa611eb02ac1bd27b97a8570cbf321437/src/libraries/System.Collections/src/System/Collections/Generic/PriorityQueue.cs#L576) method.

https://github.com/dotnet/runtime/blob/7bac4e8fa611eb02ac1bd27b97a8570cbf321437/src/libraries/System.Collections/src/System/Collections/Generic/PriorityQueue.cs#L588-L591

Contributes to https://github.com/dotnet/runtime/issues/65342