@BruceForstall 2022-04-14
You need to worry about increment/decrement overflowing a 32-bit integer and wrapping. E.g.:
```
public static int test_up_big(int[] a, int s)
    {
        int r = 0;
        int i;
        for (i = 1; i < s; i += 2147483647)
        {
            r += a[i];
        }
        return r;
    }
```
with your code will clone, but generate an access violation, not System.IndexOutOfRangeException.

In the case of `for ( ; i < x; i += c)`, `c` needs to be strictly less than `int.MaxValue - (Array.MaxLength - 1) + 1`, which is `0x7fffffff - 0x7fffffc7 + 2` = 0x3a = 58. The idea being that if `i` was the maximum allowed legal array index of 0x7fffffc6 and you added 0x3a, the result is negative (-2,147,483,648).

Similar applies to down counted loops.