@kunalspathak 2022-04-14
> public static int test_up_big(int[] a, int s)
>     {
>         int r = 0;
>         int i;
>         for (i = 1; i < s; i += 2147483647)
>         {
>             r += a[i];
>         }
>         return r;
>     }

Thanks for pointing that out. Let me try to fix that.