@kunalspathak 2022-04-19
@BruceForstall - I tried measuring some of the benchmarks that showed up in asmdiffs and they all show improvements.



| Faster                       | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| ---------------------------- | ---------:| ----------------:| ----------------:| --------:|
| BilinearTest.Interpol_Scalar |      1.23 |         12591.21 |         10245.69 |         |
| BilinearTest.Interpol_Vector |      1.08 |         27857.16 |         25908.72 |         |
| BenchmarksGame.FannkuchRedux_5.RunBench(n: 10, expectedSum: 38) |      1.06 |      16428450.00 |      15492476.47 |         |
| BenchmarksGame.ReverseComplement_1.RunBench |      1.13 |       1021397.94 |        903558.01 |         |
| ByteMark.BenchEmFloatClass |      1.05 |     771413700.00 |     734018150.00 |         |
| ByteMark.BenchEmFloat      |      1.03 |    3737632200.00 |    3641513150.00 |         |
| System.Collections.Tests.Perf_PriorityQueue<String, String>.HeapSort(Size: 100)  |     10.80 |        898710.07 |         83248.11 |         |
| System.Collections.Tests.Perf_PriorityQueue<String, String>.HeapSort(Size: 10)   |     10.44 |         39025.89 |          3739.26 |         |
| System.Collections.Tests.Perf_PriorityQueue<String, String>.HeapSort(Size: 1000) |     10.34 |      14949560.00 |       1445691.48 |         |
| System.Collections.Tests.Perf_PriorityQueue<Int32, Int32>.HeapSort(Size: 10)     |      1.04 |           203.82 |           195.96 |         |
| Benchstone.BenchI.HeapSort.Test                                                  |      1.03 |        294538.92 |        287004.51 |         |


| Slower                                                                      | diff/base | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Collections.Tests.Perf_PriorityQueue<Guid, Guid>.HeapSort(Size: 100) |      1.02 |          9744.88 |          9900.55 |         |
| System.Collections.Tests.Perf_PriorityQueue<Guid, Guid>.HeapSort(Size: 10)  |      1.01 |           396.27 |           401.45 |         |

Here is an example asmdiff from `Heapify`:

![image](https://user-images.githubusercontent.com/12488060/164083946-94249e97-1b7a-4a69-9b5b-55a2d952290b.png)